Make 'response_size' a 'size_t'

+++ tests/server.c
@@ -144,7 +144,7 @@ send_response(struct conndata *conn)
   struct rustls_connection *rconn = conn->rconn;
   const char *prefix = "HTTP/1.1 200 OK\r\nContent-Length:";
   const int body_size = 10000;
-  int response_size = strlen(prefix) + 15 + body_size;
+  size_t response_size = strlen(prefix) + 15 + body_size;
   char *response = malloc(response_size);
   size_t n;
